LIBITD-2455. Renamed "core/utils.py" to "core/logging.py"
Renamed "core/utils.py" to "core/logging.py" as that better fits the
current functionality in the file.

https://umd-dit.atlassian.net/browse/LIBITD-2455
