+++ src/textbooks/web.py
@@ -1,6 +1,6 @@
 from typing import Any, Optional, Text, TextIO
 
-from core.utils import create_logger
+from core.logging import create_logger
 from core.web_errors import blueprint
 from flask import Flask, abort, request
 from yaml import safe_load
